Created source directory and moved the two classes into it
